+++ src/steps/users/index.ts
@@ -1,0 +1,58 @@
+import {
+  createDirectRelationship,
+  Entity,
+  IntegrationStep,
+  IntegrationStepExecutionContext,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
+
+import { createAPIClient } from '../../client';
+import { IntegrationConfig } from '../../config';
+import {
+  Entities,
+  Relationships,
+  Steps,
+  ACCOUNT_ENTITY_KEY,
+} from '../constants';
+import { createUserEntity } from './converter';
+
+export async function fetchUsers({
+  logger,
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+  const accountEntity = (await jobState.getData(ACCOUNT_ENTITY_KEY)) as Entity;
+
+  try {
+    await apiClient.iterateUsers(async (user) => {
+      const userEntity = await jobState.addEntity(createUserEntity(user));
+      await jobState.addRelationship(
+        createDirectRelationship({
+          _class: RelationshipClass.HAS,
+          from: accountEntity,
+          to: userEntity,
+        }),
+      );
+    });
+  } catch (err) {
+    if (err.status === 403) {
+      logger.publishEvent({
+        name: 'missing_permission',
+        description:
+          '"User.Read.All" is a a required permission to run the Microsoft Teams integration, it is required for getting the user\'s data.',
+      });
+    }
+  }
+}
+
+export const userSteps: IntegrationStep<IntegrationConfig>[] = [
+  {
+    id: Steps.USERS,
+    name: 'Fetch Users',
+    entities: [Entities.USER],
+    relationships: [Relationships.ACCOUNT_HAS_USER],
+    dependsOn: [Steps.ACCOUNT],
+    executionHandler: fetchUsers,
+  },
+];